default D3 graphing
