--- packages/messier-61-graph/D3Graph/ConfigurableD3Graph.module.css
@@ -1,64 +1,0 @@
-/*
- * Copyright Jiaqi Liu
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-marker#child {
-  fill: rgb(255, 120, 25);
-}
-
-svg.g.node circle {
-  fill: rgb(88, 155, 213);
-  stroke: #fff;
-  stroke-width: 1.2px;
-}
-
-svg.g.node text {
-  fill: black;
-}
-
-g.node.selectedSource circle {
-  fill: rgb(30, 197, 55);
-  stroke-width: 1.2px;
-}
-
-g.node.selectedTarget circle {
-  fill: rgb(30, 197, 55);
-  stroke-width: 1.2px;
-}
-
-svg.g.node a {
-  pointer-events: none;
-  font: 13px sans-serif;
-}
-
-svg.g.node a text {
-  stroke: blue;
-}
-
-svg.line.link {
-  stroke: rgb(255, 120, 25);
-  /* stroke-opacity: .8; */
-  stroke-width: 1.2px;
-}
-
-svg.line.newLine {
-  stroke: rgb(30, 197, 55);
-  /* stroke-opacity: .8; */
-  stroke-width: 1.2px;
-}
-
-line.selected {
-  stroke: rgb(58, 204, 80);
-  stroke-opacity: 1;
-}
