+++ packages/messier-61-graph/D3Graph/D3Graph.module.css
@@ -1,0 +1,53 @@
+marker#child {
+  fill: rgb(255, 120, 25);
+}
+
+svg.g.node circle {
+  fill: rgb(88, 155, 213);
+  stroke: #fff;
+  stroke-width: 1.2px;
+}
+
+svg.g.node text {
+  fill: black;
+}
+
+g.node.selectedSource circle {
+  fill: rgb(30, 197, 55);
+  stroke-width: 1.2px;
+}
+
+g.node.selectedTarget circle {
+  fill: rgb(30, 197, 55);
+  stroke-width: 1.2px;
+}
+
+svg.g.node a {
+  pointer-events: none;
+  font: 13px sans-serif;
+}
+
+svg.line.newLine {
+  stroke: rgb(30, 197, 55);
+  /* stroke-opacity: .8; */
+  stroke-width: 1.2px;
+}
+svg.line.link {
+  stroke: rgb(255, 120, 25);
+  /* stroke-opacity: .8; */
+  stroke-width: 1.2px;
+}
+
+svg.g.node text {
+  stroke: black;
+  font: 13px sans-serif;
+}
+
+line.selected {
+  stroke: rgb(58, 204, 80);
+  stroke-opacity: 1;
+}
+
+rect {
+  fill: white;
+}
